refactor escape_error_output test
